Made All The Other Pages and Linked Them
Also the header thing you made is weird.. because now all headers will
be white with a blue background which, i don't know maybe you actually
wanted this.

To make normal headers I had to make a class in the .css file called
normalHeader.. this might become an issue, to fix it I think we should
make the blue/white header a seperate special class so then the default
is the default and our lives are easier.